requirements.txt
# Core dependencies
requests>=2.25.0
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Platform-specific APIs
praw>=7.4.0  # Reddit API
google-api-python-client>=2.0.0  # YouTube API
feedparser>=6.0.0  # RSS feeds

# OpenAI integration
openai>=1.0.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0

# Optional: For enhanced functionality
beautifulsoup4>=4.9.0  # HTML parsing
lxml>=4.6.0  # XML parsing
selenium>=4.0.0  # Web scraping (if needed)

# Development dependencies
pytest>=6.0.0  # Testing
black>=21.0.0  # Code formatting
flake8>=3.8.0  # Linting

